+++ src/main/java/com/jaquadro/minecraft/storagedrawers/api/storage/BlockType.java
@@ -1,6 +1,0 @@
-package com.jaquadro.minecraft.storagedrawers.api.storage;
-
-public enum BlockType {
-    Drawers,
-    Trim
-}