+++ src/main/java/com/jaquadro/minecraft/storagedrawers/api/storage/EmptyDrawerAttributes.java
@@ -1,6 +1,0 @@
-package com.jaquadro.minecraft.storagedrawers.api.storage;
-
-public class EmptyDrawerAttributes implements IDrawerAttributes {
-    public EmptyDrawerAttributes() {
-    }
-}