+++ src/main/java/com/jaquadro/minecraft/storagedrawers/api/storage/INetworked.java
@@ -1,4 +1,0 @@
-package com.jaquadro.minecraft.storagedrawers.api.storage;
-
-public interface INetworked {
-}