+++ src/main/java/com/jaquadro/minecraft/storagedrawers/api/storage/attribute/IVoidable.java
@@ -1,7 +1,0 @@
-package com.jaquadro.minecraft.storagedrawers.api.storage.attribute;
-
-@Deprecated
-public interface IVoidable {
-    @Deprecated
-    boolean isVoid();
-}